Shop Score Float 처리

+++ database/migrations/2021_11_10_083332_create_shops_table.php
@@ -26,7 +26,7 @@ class CreateShopsTable extends Migration
 //            $table->string('영업일')->nullable();
             $table->string('address');
             $table->string('address2')->nullable();
-            $table->integer('score')->default(0);
+            $table->float('score')->default(0);
             $table->integer('score_total')->default(0);
             $table->integer('score_count')->default(0);
             $table->point('location',4326); // google map 3857
